refactor pasteFromClipboard
